Tutorial submission functions

The following function was modified from Colin Rundel's learnrhash package, available at https://github.com/rundel/learnrhash.

This note is also modified from Professor Rundel's description: Note that when including these functions in a learnr Rmd document it is necessary that the server function, submission_server(), be included in an R chunk where context="server". Conversely, any of the ui functions, *_ui(), must not be included in an R chunk with a context.

submission_server(session)

submission_ui

Arguments

session
session object from shiny with learnr

Format

An object of class shiny.tag of length 3.